Expose getNetworkingAllowedForUid for netd
A piece of "Disallow creating network sockets if uid is restricted",
ported to 13.

This patch aims to address the separation of netd functionality into
the Connectivity module, and the challenge of allowing netd itself to
check whether networking is allowed for a given uid, now that
TrafficController has been moved to this module instead.

It exposes a new function libnetd_updatable_getNetworkingAllowedForUid
for use by netd in determining whether a uid should be allowed to
create a network socket. getNetworkingAllowedForUid is ultimately
implemented in netd/BpfHandler.cpp and is nearly identical to
Oliver Scott's function previously found in netd's TrafficController.

Changing libnetd_updatable requires updating the abi dumps:
`development/vndk/tools/header-checker/utils/create_reference_dumps.py
-l libnetd_updatable`

This updates prebuilts/abi-dumps/platform, which should probably be
forked.

If abi dumps are not updated, the following error will occur:
> error: VNDK library: libnetd_updatable's ABI has EXTENDING CHANGES
> Please check compatibility report at: out/soong/.intermediates/
> packages/modules/Connectivity/netd/libnetd_updatable/
> android_x86_64_shared_cfi_apex30/libnetd_updatable.so.abidiff

Finally, the definition of uid_owner_map in bpf_progs/netd.c is
updated so that selinux policies will allow netd read-only access to
the maps. Otherwise, libnetd_updatable will fail to initialize, and
the system will fail to boot.

Issue: calyxos#581
Co-authored-by: Oliver Scott <olivercscott@gmail.com>
Change-Id: I092d187cd25c2c48708b2a22b7a7a7a732b87afb