Always drop non-VPN ingress in VPN lockdown mode
When "Block connections without VPN" is specified, incoming traffic
from non-VPN interfaces should be blocked, even for VPNs that are not
fully-routed. Outgoing traffic to non-VPN interfaces is already blocked
in this situation; allowing incoming but not outgoing makes no sense.
(Loopback is excluded as usual.)

We accomplish this by always setting the ingress interface in the BPF
UidOwnerMap for apps covered by a VPN, but we only consider it when
the IIF_MATCH or (new logic) LOCKDOWN_VPN_MATCH rule are specified.

Issue: calyxos#1255
Change-Id: Id7954816566cb06bf2e9869ea98b20678835df9d