Always drop non-VPN ingress in VPN lockdown mode
When "Block connections without VPN" is specified, incoming traffic
from non-VPN interfaces should be blocked, even for VPNs that are not
fully-routed or otherwise not considered eligible for isolation.
Outgoing traffic to non-VPN interfaces is already blocked in this
situation. (Loopback is excluded as usual.)

We accomplish this by altering getVpnIsolationInterface such that the
VPN interface is always returned in lockdown mode, and ensuring that
the allowed interface rules are updated when the lockdown mode changes.

Co-authored-by: Oliver Scott <olivercscott@gmail.com>
Issue: calyxos#1255
Change-Id: Id7954816566cb06bf2e9869ea98b20678835df9d