Always drop non-VPN ingress in VPN lockdown mode
When "Block connections without VPN" is specified, incoming traffic
from non-VPN interfaces should be blocked, even for VPNs that are not
fully-routed. Outgoing traffic to non-VPN interfaces is already blocked
in this situation. (Loopback is excluded as usual.)

We accomplish this by ensuring the allowed interface for BPF's
mUidOwnerMap is set to the actual VPN interface when lockdown is in
effect for a uid. We still supply the null interface to the maps when
appropriate, but we track allowed-ingress uids separately so as to
retain knowledge of their actual VPN interface, which must be known
in order to update the uid interface rules appropriately.

Co-authored-by: Oliver Scott <olivercscott@gmail.com>
Issue: calyxos#1255
Change-Id: Id7954816566cb06bf2e9869ea98b20678835df9d