Accommodate missing LineageSettings in tests
VPN tests rely on methods that involve LineageSettings.
Call a newly-added constructor to force a given result of checks
for whether Global VPN is turned on. Perform tests with Global VPN
assumed to be turned off, for now. We have never had our own unit
tests related to Global VPN, although we may add some in the future.

Change-Id: I82bdd81755f94ef2c616aee053446b9c9fec78a6