Set allowed UIDs for networks based on policies
All networks are restricted, and only UIDs whose policy allows them
on a transport are able to route outgoing connections over a network
of that type. In addition, to partially address incoming traffic, and
to handle UIDs that have a policy which preevents them from accessing
VPNs, UIDs that are either not allowed to access the VPN that applies
to them or are not allowed to access the default network are added to
a denylist that prevents them from accessing networks in general.

Change-Id: I79342edbec92090cca20853ba50ea7fd48ec81c2