Add denylist listener and failure fallback
Add a denylist changed listener for use by NetworkPolicyManagerService
to be notified when a UID is blocked from accessing all networks based
on an allowed-transport determination.

Add a fallback for UIDs possibly failing to be adding to or removed
from the denylist, replacing the chain instead in such an event. It is
not clear what would cause such an error; this is just in case.

Change-Id: I7e3beaf8fca45bfe52737051a6704b4c52791a7b