DO NOT MERGE: Test firewall without denylist
With this change, the "allowed UIDs" component of the firewall can be
tested independently of the denylist, which is necessary to ensure that
network transition-related leaks are not possible, because the routing
rules are not present.

A consequence of this change is that the VPN firewall toggle has no
effect, as this toggle entirely depends on the denylist functionality.

Change-Id: Id90a1781f2ed5e7abb4078545156439cdbf73fa5