Use framework listener for allowed transports
When allowed transports change, instead of exposing an API from
Connectivity that the framework calls, add a new method to the
framework's NetworkPolicyCallback that we override.

This change benefits the prebuilt mainline module, preventing
the need to fight with the platform to add new module APIs.

Requires fw/b change of the same Change-Id.

Squash with: I79342edbec92090cca20853ba50ea7fd48ec81c2
"Set allowed UIDs for networks based on policies"

Change-Id: Ie476f23684b00397197184e965201d6823b28de2