Notify the framework about denylist changes
When the denylist changes, instead of exposing an API from
Connectivity that the framework calls, add a new method to the
framework's NetworkPolicyManager that we call to inform it of
this directly.

This change benefits mainline modules, preventing the need to fight
the platform to add new module APIs.

Requires fw/b change of the same Change-Id.

Squash with: I2729b61c349ec2812a74d7d1c04b90a58b0f5b88
"Expose new isUidCurrentlyDisallowedByPolicy"

Change-Id: I3c3593f110753a3ce02af3739f600190f22e9663